+++ fltk/src/lib.rs
@@ -24,6 +24,7 @@ Why choose FLTK?
 
 Resources:
 - [Book](https://fltk-rs.github.io/fltk-book/)
+- [本書的中文翻譯](https://flatig.vip/fltk-book-zh)
 - [Documentation](https://docs.rs/fltk)
 - [Videos](https://github.com/fltk-rs/fltk-rs#tutorials)
 - [Discussions](https://github.com/fltk-rs/fltk-rs/discussions)